+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## Unreleased
+
+## [v0.5.6]
+> December 14, 2016
+
 - **Minor** ADd `Disposable` and rework Resource disposable handling
 - **Feature** Add `Mutable` helepr methods `unsetErrors`, `setErrors`, `isMutated`
+
+[v0.5.6]: https://github.com/netarc/refrax/compare/v0.5.5...v0.5.6
 
 ## [v0.5.5]
 > November 17, 2016